pyproject.toml
[project]
name = "clockiduck"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Henrik Gregersen", email = "hgr@alfapeople.com" }
]
requires-python = ">=3.13"
dependencies = [
    "requests",
    "pydantic-settings",
    "duckdb",
]

[project.scripts]
clociduck-sync = "clockiduck.sync:sync_clockify_to_db"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
